成交量字段因数据过大，更换成double型

+++ yuna/destinations/mysql_peewee.py
@@ -74,5 +74,5 @@ class Details(Model):
     low = DoubleField()
     high = DoubleField()
     close = DoubleField()
-    volume = IntegerField()
+    volume = DoubleField()
